Implement dockerized Keras for model training
Keras with a TensorFlow backend is used to train the QSAR dataset.
The Sequential model architecture was created based on the Ma et al.
paper. Keras was install via Docker container to make the training
script platform agnostic. Furthermore, pandas library was added to the
docker build image process and was removed from the Pipfile.